backend/requirements.txt
# === CORE ===
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# === DATABASE ===
sqlalchemy==2.0.23
alembic==1.13.1

# === SECURITY ===
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
pydantic==2.5.1
pydantic-settings==2.1.0

# === VALIDATION ===
email-validator==2.1.0

# === ML & IMAGE PROCESSING ===
tensorflow==2.14.0
keras==2.14.0
numpy==1.26.2
opencv-python==4.8.1.78
pillow==10.1.0
scikit-image==0.22.0

# === LLM ===
openai==1.3.9
langchain==0.1.1

# === UTILITIES ===
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1

# === MONITORING & LOGGING ===
structlog==23.3.0
python-json-logger==2.0.7

# === TESTING ===
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-cov==4.1.0

# === CODE QUALITY ===
black==23.12.0
flake8==6.1.0
pylint==3.0.3
isort==5.13.2
mypy==1.7.1

# === DOCUMENTATION ===
mkdocs==1.5.3
mkdocs-material==9.5.3

# === DEVELOPMENT ===
ipython==8.18.1
